refactor: Moved the "combinedPopulationByYearData" calculation into the "LineGraphComponent". And  use "FC type"
