+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import type { AxiosInstance } from 'axios';
 import axios from 'axios';
-import React, { useEffect, useMemo, useState } from 'react';
+import type { FC } from 'react';
+import { useEffect, useState } from 'react';
 import LineGraphComponent from '../components/LineGraphComponent/LineGraph';
 import PrefectureCheckBoxes from '../components/PrefectureCheckBoxes/PrefectureCheckBoxes';
 import type {
@@ -18,7 +19,7 @@ const resasAxiosInstance: AxiosInstance = axios.create({
   },
 });
 
-const Home: React.FC = () => {
+const Home: FC = () => {
   const [prefecturesWithCheck, setPrefecturesWithCheck] = useState<PrefectureWithCheck[]>([]);
   const [totalPopulationData, setTotalPopulationData] = useState<PopulationData[]>([]);
 
@@ -84,29 +85,6 @@ const Home: React.FC = () => {
     }
   };
 
-  const combinedPopulationByYearData = useMemo(() => {
-    const yearMap: { [key: number]: { year: number; [key: string]: number | null } } = {};
-
-    totalPopulationData.forEach((popData) => {
-      popData.data.forEach((entry) => {
-        if (yearMap[entry.year] === undefined) {
-          yearMap[entry.year] = { year: entry.year };
-        }
-        yearMap[entry.year][popData.prefName] = entry.value;
-      });
-    });
-
-    Object.values(yearMap).forEach((yearData) => {
-      totalPopulationData.forEach((popData) => {
-        if (!(popData.prefName in yearData)) {
-          yearData[popData.prefName] = null;
-        }
-      });
-    });
-
-    return Object.values(yearMap);
-  }, [totalPopulationData]);
-
   useEffect(() => {
     fetchPrefectures();
   }, []);
@@ -119,7 +97,6 @@ const Home: React.FC = () => {
       />
 
       <LineGraphComponent
-        combinedData={combinedPopulationByYearData}
         populationData={totalPopulationData}
         prefecturesWithCheck={prefecturesWithCheck}
       />
